Refactor EpisodeDetails: Move DanmakuEditor to `.ui.subject.episode.danmaku`
